requirements.txt
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
mysqlclient==1.4.5
Pillow==6.2.1
protobuf==3.10.0
pycparser==2.19
six==1.12.0
SQLAlchemy==1.3.8
Werkzeug==0.16.0
WTForms==2.2.1
